backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
Pillow==11.3.0
django-mongodb-backend==5.2.0b2
psycopg2-binary==2.9.10
google-generativeai==0.8.5

